style: 주석 삭제 및 경로 변경
