+++ src/api/category/index.ts
@@ -1,6 +1,12 @@
 import { fetcher } from "api/fetcher";
-import { CategoryType } from "./types";
+// import { CategoryType } from "./types";
 import { Response } from "api/types";
+
+export type CategoryType = {
+  id: number;
+  title: string;
+  imageUrl: string;
+};
 
 export const getCategories = async (): Promise<Response<CategoryType[]>> => {
   const res = await fetcher.get("/categories");
