--- src/api/category/types.ts
@@ -1,5 +1,0 @@
-export type CategoryType = {
-  id: number;
-  title: string;
-  imageUrl: string;
-};
