+++ src/api/region/index.ts
@@ -1,6 +1,11 @@
 import { fetcher } from "api/fetcher";
-import { RegionType } from "./types";
+
 import { Response } from "api/types";
+
+export type RegionType = {
+  id: number;
+  title: string;
+};
 
 export const getRegionList = async (): Promise<Response<RegionType[]>> => {
   const res = await fetcher.get("/regions");
