--- src/api/region/types.ts
@@ -1,4 +1,0 @@
-export type RegionType = {
-  id: number;
-  title: string;
-};
