+++ src/api/user/index.ts
@@ -1,5 +1,10 @@
 import { fetcher } from "../fetcher";
 
+export type User = {
+  nickname: string;
+  profileUrl: string;
+};
+
 export const postSignUp = async (body: FormData) => {
   return await fetcher.post<null>("/users/signup", body);
 };
