--- src/api/user/types.ts
@@ -1,5 +1,0 @@
-export type User = {
-  nickname: string;
-  profileUrl: string;
-};
-
