+++ src/components/Category/CategoryItem.tsx
@@ -1,5 +1,6 @@
 import { ModalListItem } from "@components/common/Modal/ModalStyles";
-import { CategoryType } from "api/category/types";
+import { CategoryType } from "api/category";
+
 
 type Props = {
   item: CategoryType;
