+++ src/router/ProtectedRoute.tsx
@@ -1,15 +1,14 @@
-import useUser from "api/queries/useUser";
-import { useEffect } from "react";
-import { Outlet, useNavigate } from "react-router-dom";
+// import { useEffect } from "react";
+import { Outlet } from "react-router-dom";
 
 export default function ProtectedRoute() {
   // TODO: replace with auth context.
-  const { data: user, isFetched } = useUser();
-  const navigate = useNavigate();
+  // const { data: user, isFetched } = useUser();
+  // const navigate = useNavigate();
 
-  useEffect(() => {
-    if (isFetched && !user) navigate("/signin");
-  }, [isFetched, user, navigate]);
+  // useEffect(() => {
+  //   if (isFetched && !user) navigate("/signin");
+  // }, [isFetched, user, navigate]);
 
   return <Outlet />;
 }
